requirements.txt
asttokens==3.0.0
bottle==0.13.3
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
proxy_tools==0.1.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.17
python-dateutil==2.9.0.post0
pywebview==5.4
pyzmq==26.4.0
QtPy==2.4.3
six==1.17.0
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.0
wcwidth==0.2.13
